third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/set-outer-html-body-expected.txt

Tests DOMAgent.setOuterHTML invoked on body tag. See https://bugs.webkit.org/show_bug.cgi?id=62272. 


Running: testSetUp

Running: testSetBody
<html id="html"><head id="head">
<script>
</script>
</head>
<body>New body content</body></html>

Running: testInsertComments
<html id="html"><head id="head">
<script>
</script>
</head>
<!-- new comment between head and body --><body>New body content</body></html>

Running: testSetHead
<html id="html"><head><!-- new head content --></head>
<!-- new comment between head and body --><body>New body content</body></html>

Running: testSetHTML
<html><head><!-- new head content --></head><body>Setting body as a part of HTML.</body></html>
